rename directory
https://coveord.atlassian.net/browse/KIT-466
