+++ packages/headless/src/controllers/history-manager/headless-history.ts
