+++ packages/headless/src/controllers/index.ts
@@ -72,13 +72,17 @@ export {
   NumericFacet,
 } from './facets/range-facet/numeric-facet/headless-numeric-facet';
 
-export {History, HistoryState, buildHistory} from './history/headless-history';
+export {
+  History,
+  HistoryState,
+  buildHistory,
+} from './history-manager/headless-history';
 
 export {
   HistoryManager,
   HistoryManagerState,
   buildHistoryManager,
-} from './history/headless-history-manager';
+} from './history-manager/headless-history-manager';
 
 export {
   PagerInitialState,
